Remove Channel and ChannelCounterparty traits
