+++ modules/src/ics04_channel/mod.rs
@@ -3,6 +3,5 @@
 pub mod channel;
 pub mod error;
 pub mod events;
-pub mod exported;
 pub mod msgs;
 pub mod packet;
