+++ relayer-cli/tests/integration.rs
@@ -14,10 +14,7 @@
 use ibc::ics03_connection::connection::ConnectionEnd;
 use ibc::ics03_connection::exported::Connection;
 use ibc::ics03_connection::exported::State as ConnectionState;
-use ibc::ics04_channel::channel::ChannelEnd;
-use ibc::ics04_channel::exported::Channel;
-use ibc::ics04_channel::exported::Order;
-use ibc::ics04_channel::exported::State as ChannelState;
+use ibc::ics04_channel::channel::{ChannelEnd, Order, State as ChannelState};
 use ibc::ics23_commitment::CommitmentPrefix;
 use ibc::ics24_host::identifier::{ChannelId, ClientId, ConnectionId, PortId};
 use ibc::ics24_host::Path::{ChannelEnds, ClientConnections, Connections};
